* move phpunit tests from `be` to `symfony`
